Renamed ImageViewer to LabelImageLoader
